Init auth lib
