+++ auth/dataplane/dataplane_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/databricks/databricks-sdk-go/databricks/config/experimental/auth"
+	"github.com/databricks/databricks-sdk-go/auth"
 	"golang.org/x/oauth2"
 )
