+++ auth/go.mod
@@ -1,0 +1,5 @@
+module github.com/databricks/databricks-sdk-go/auth
+
+go 1.18
+
+require golang.org/x/oauth2 v0.24.0
